Split out code to build C argv
